Further cleanup of followers
